+++ package.json
@@ -5,8 +5,8 @@
   "homepage": "https://molchanovboris.github.io/carsharing_react_molchanov_boris/",
   "dependencies": {
     "antd": "^4.14.0",
+    "eslint": "^7.22.0",
     "gh-pages": "^3.1.0",
-    "node-sass": "^5.0.0",
     "react": "^17.0.1",
     "react-bootstrap": "^1.5.2",
     "react-dom": "^17.0.1",
@@ -40,5 +40,9 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "node-sass": "^5.0.0",
+    "prettier": "2.2.1"
   }
 }
