+++ src/components/Info/index.js
@@ -1,0 +1,3 @@
+import Info from "./Info";
+
+export default Info;
