+++ src/components/Slider/slides.scss
@@ -53,18 +53,44 @@
 
   .carousel-inner {
     .slider {
-      .slider-image {
+      &__image {
         height: 100vh;
         width: 100%;
       }
 
-      .slider-gradient {
+      &__gradient {
         top: 0;
         left: 0;
         position: absolute;
         width: 100%;
         height: 100%;
         background: linear-gradient(180deg, rgba(0, 0, 0, 0) 0%, #000000 100%);
+      }
+
+      &__title {
+        margin-bottom: 8px;
+        font-size: 48px;
+        font-weight: 500;
+        line-height: 47px;
+        color: $white-color1;
+      }
+
+      &__subtitle {
+        margin-bottom: 32px;
+        font-size: 24px;
+        font-weight: 300;
+        line-height: 90%;
+        color: $white-color2;
+      }
+
+      &__button {
+        height: 48px;
+        width: 164px;
+        border-radius: 4px;
+        border: none;
+        font-size: 18px;
+        line-height: 21px;
+        color: $white-color2;
       }
 
       .carousel-caption {
@@ -74,32 +100,6 @@
         flex-direction: column;
         justify-content: center;
         text-align: left;
-
-        .slider-title {
-          margin-bottom: 8px;
-          font-size: 48px;
-          font-weight: 500;
-          line-height: 47px;
-          color: $white-color1;
-        }
-
-        .slider-sub-title {
-          margin-bottom: 32px;
-          font-size: 24px;
-          font-weight: 300;
-          line-height: 90%;
-          color: $white-color2;
-        }
-
-        .slider-button {
-          height: 48px;
-          width: 164px;
-          border-radius: 4px;
-          border: none;
-          font-size: 18px;
-          line-height: 21px;
-          color: $white-color2;
-        }
       }
     }
   }
