+++ src/components/sidebar/sidebar.scss
@@ -1,0 +1,111 @@
+@import "../../assets/styles/color.scss";
+
+.sidebar {
+  width: 64px;
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  background-color: $black-color2;
+
+  @media screen and (max-width: 780px) {
+    width: 86px;
+  }
+
+  @media screen and (max-width: 450px) {
+    position: absolute;
+    height: 50px;
+    background-color: $white-color1;
+  }
+
+  &__show,
+  &__language {
+    margin: 40px 0 31px 0;
+    border: none;
+    background: $black-color2 !important;
+
+    @media screen and (max-width: 450px) {
+      display: none;
+    }
+  }
+
+  &__language {
+    font-style: normal;
+    font-size: 13px;
+    line-height: 15px;
+    color: $green-color;
+    text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  }
+}
+
+.drawer {
+  .ant-drawer-mask {
+    background-color: rgba(21, 27, 31, 0.81);
+  }
+
+  &__close.ant-btn {
+    position: absolute;
+    padding: 40px 24px;
+    border: none;
+    background: $black-color3;
+
+    @media screen and (max-width: 1040px) {
+      padding: 24px;
+    }
+  }
+
+  &__block {
+    margin-left: 128px;
+    display: flex;
+    flex-direction: column;
+    flex: 1;
+    justify-content: center;
+    text-transform: uppercase;
+
+    @media screen and (max-width: 1040px) {
+      margin-left: 28px;
+      margin-top: 80px;
+      justify-content: start;
+    }
+  }
+
+  &__link {
+    margin-bottom: 38px;
+    font-style: normal;
+    font-weight: 500;
+    font-size: 32px;
+    line-height: 37px;
+    color: $white-color1;
+  }
+
+  &__link:focus,
+  &__link:hover {
+    color: $green-color;
+  }
+
+  &__socials {
+    margin-right: 24px;
+  }
+
+  .ant-drawer-content-wrapper {
+    width: 55% !important;
+
+    @media screen and (max-width: 1040px) {
+      width: 65% !important;
+    }
+
+    @media screen and (max-width: 1040px) {
+      width: 100% !important;
+    }
+
+    .ant-drawer-content {
+      background-color: $black-color3;
+
+      .ant-drawer-wrapper-body .ant-drawer-body {
+        padding: 0;
+        display: flex;
+        flex-direction: column;
+      }
+    }
+  }
+}
