--- src/components/sidebar/style.scss
@@ -1,123 +1,0 @@
-@import "../../assets/styles/color.scss";
-
-.sidebar {
-  width: 64px;
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  background-color: $black-color2;
-
-  @media screen and (max-width: 780px) {
-    width: 86px;
-  }
-
-  @media screen and (max-width: 450px) {
-    position: absolute;
-    height: 50px;
-    background-color: $white-color1;
-  }
-
-  .ant-btn.sidebar-show-button,
-  .ant-btn.sidebar-language {
-    margin: 40px 0 31px 0;
-    border: none;
-    background: $black-color2 !important;
-
-    @media screen and (max-width: 450px) {
-      display: none;
-    }
-  }
-
-  .ant-btn.sidebar-show-small-button {
-    display: none;
-
-    @media screen and (max-width: 450px) {
-      margin: 16px;
-      padding: 0;
-      display: flex;
-      border: none;
-      background-color: $white-color1;
-    }
-  }
-
-  .sidebar-language {
-    font-style: normal;
-    font-size: 13px;
-    line-height: 15px;
-    color: $green-color;
-    text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-  }
-}
-
-.ant-drawer {
-  .ant-drawer-mask {
-    background-color: rgba(21, 27, 31, 0.81);
-  }
-
-  .ant-drawer-content-wrapper {
-    width: 55% !important;
-
-    @media screen and (max-width: 1040px) {
-      width: 65% !important;
-    }
-
-    @media screen and (max-width: 1040px) {
-      width: 100% !important;
-    }
-
-    .ant-drawer-content {
-      background-color: $black-color3;
-
-      .ant-drawer-wrapper-body .ant-drawer-body {
-        padding: 0;
-        display: flex;
-        flex-direction: column;
-
-        .ant-btn {
-          position: absolute;
-          padding: 40px 24px;
-          border: none;
-          background: $black-color3;
-
-          @media screen and (max-width: 1040px) {
-            padding: 24px;
-          }
-        }
-
-        .sidebar-block {
-          margin-left: 128px;
-          display: flex;
-          flex-direction: column;
-          flex: 1;
-          justify-content: center;
-          text-transform: uppercase;
-
-          @media screen and (max-width: 1040px) {
-            margin-left: 28px;
-            margin-top: 80px;
-            justify-content: start;
-          }
-
-          .sidebar-items {
-            margin-bottom: 38px;
-            font-style: normal;
-            font-weight: 500;
-            font-size: 32px;
-            line-height: 37px;
-            color: $white-color1;
-          }
-
-          .sidebar-items:focus,
-          .sidebar-items:hover {
-            color: $green-color;
-          }
-
-          .sidebar-society-links .sidebar-society-item {
-            margin-right: 24px;
-          }
-        }
-      }
-    }
-  }
-}
