+++ src/pages/MainPage/style.scss
